pyproject.toml
[tool.poetry]
name = "crael"
version = "0.1.1"
description = "A tormenta RPG companion bot"
authors = ["Andre Heringer <andregomesh@dcc.ufmg.br>"]

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.2"
isort = {version = "^4.3", extras = ["pyproject"]}

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
black = {version = "^18.3-alpha.0"}
flake8-docstrings = "^1.3"
xdoctest = "^0.10.0"
pytest-cov = "^2.7"
pytest-mock = "^1.10"
coverage = "^4.5"
tox = "^3.13"
pytest = "^5.1"
pytest-asyncio = "^0.10.0"
pre-commit = "^1.18"
sphinx = "^2.2"
sphinx_rtd_theme = "^0.4.3"
darglint = "^0.6.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
